+++ R/zzz-aliases.R
@@ -1,0 +1,36 @@
+#' Alias for `style_tt()`
+#'
+#' @rdname style_tt
+#' @export
+tt_style <- style_tt
+
+#' Alias for `group_tt()`
+#'
+#' @rdname group_tt
+#' @export
+tt_group <- group_tt
+
+#' Alias for `format_tt()`
+#'
+#' @rdname format_tt
+#' @export
+tt_format <- format_tt
+
+#' Alias for `plot_tt()`
+#'
+#' @rdname plot_tt
+#' @export
+tt_plot <- plot_tt
+
+#' Alias for `strip_tt()`
+#'
+#' @rdname strip_tt
+#' @export
+tt_strip <- strip_tt
+
+
+#' Alias for `save_tt()`
+#'
+#' @rdname save_tt
+#' @export
+tt_save <- save_tt