+++ NEWS.md
@@ -1,5 +1,7 @@
 # News
 
+## Development version
+
 ## 0.12.0
 
 Deprecated or replaced features: